requirements.txt
apps==0.0.0.0.dev20200918
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.4
defusedxml==0.7.0rc2
dj-rest-auth==2.1.3
Django==3.0.7
django-allauth==0.44.0
django-appconf==1.0.4
django-braces==1.14.0
django-common-helpers==0.9.2
django-compressor==2.4
django-cors-headers==3.7.0
django-cron==0.5.1
django-crontab==0.7.1
django-jsonfield==1.4.1
django-libsass==0.8
django-nested-admin==3.2.2
django-oauth-toolkit==1.4.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-rest-passwordreset==1.1.0
django-reversion==3.0.9
django-templated-mail==1.1.1
django-tinymce==3.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gphotospy==0.1.3
httplib2==0.19.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
libsass==0.20.1
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.0
pkg-resources==0.0.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2021.1
rcssmin==1.0.6
requests==2.25.1
requests-oauthlib==1.3.0
rest-social-auth==5.0.1
rjsmin==1.1.0
rsa==4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.4.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
